// src/app/models/scholarship-request.model.ts
export interface ScholarshipRequestModel {
    firstName: string;
    lastName: string;
    birthDate: string;
    gender: string;
    street: string;
    city: string;
    state: string;
    zip: string;
    email: string;
    phone: string;
    parentName: string;
    timePlayed: string;
    clubHistory: string;
    positionsPlayed: string;
    tournamentHistory: string;
    whySoccer: string;
    householdIncome: string;
    householdSize: string;
    financialChallenges: string;
}